Unify .NET Framework/Core code paths
CelesteNet now builds against .NET Core Everest, meaning .NET Framework
code paths will never be taken, and .NET Core paths always will.

This commit also removes the NETFRAMEWORK/NETCORE define constants
and .NET Framework exclusive references, as they're now unused.
