+++ CelesteNet.Server.FrontendModule/CelesteNet.Server.FrontendModule.csproj
@@ -14,30 +14,18 @@
 
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+
+    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
+    <PackageReference Include="System.Drawing.Common" Version="8.0.2" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.8.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
+
     <!--
       this is a build of [websocket-sharp with custom headers](https://github.com/felixhao28/websocket-sharp)
       i just had to merge main into it (since it was severely out of date) and converted the .csproj to build against .net 7
       -->
     <Reference Include="websocket-sharp-customheaders.dll" HintPath="..\lib\websocket-sharp-customheaders.dll" />
   </ItemGroup>
-
-  <Choose>
-    <When Condition="$(TargetFramework.Contains('.'))">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-        <PackageReference Include="System.Drawing.Common" Version="8.0.2" />
-        <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.8.0" />
-        <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
-      </ItemGroup>
-    </When>
-
-    <Otherwise>
-      <ItemGroup>
-        <Reference Include="System.Web" />
-        <Reference Include="System.Drawing" />
-      </ItemGroup>
-    </Otherwise>
-  </Choose>
 
 </Project>
