+++ CelesteNet.Server/CelesteNet.Server.csproj
@@ -30,31 +30,16 @@
 
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-  </ItemGroup>
 
-  <Choose>
-    <When Condition="$(TargetFramework.Contains('.'))">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-        <PackageReference Include="System.Drawing.Common" Version="8.0.2" />
-        <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.8.0" />
-        <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
-      </ItemGroup>
-    </When>
-
-    <Otherwise>
-      <ItemGroup>
-        <Reference Include="System.Web" />
-        <Reference Include="System.Drawing" />
-      </ItemGroup>
-    </Otherwise>
-  </Choose>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="8.0.2" Condition="$(TargetFramework.Contains('.'))" />
-    <PackageReference Include="MessagePack" Version="2.5.140" Condition="$(TargetFramework.Contains('.'))" />
-    <PackageReference Include="MessagePackAnalyzer" Version="2.5.140" Condition="$(TargetFramework.Contains('.'))">
+    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
+    <PackageReference Include="System.Drawing.Common" Version="8.0.2" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.8.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
+      
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="8.0.2" />
+    <PackageReference Include="MessagePack" Version="2.5.140" />
+    <PackageReference Include="MessagePackAnalyzer" Version="2.5.140">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
